refactor: remove `global` from facade/lang
