+++ modules/@angular/core/src/core_private_export.ts
@@ -23,5 +23,6 @@ export {ReflectionCapabilities as ɵReflectionCapabilities} from './reflection/r
 export {ReflectorReader as ɵReflectorReader} from './reflection/reflector_reader';
 export {GetterFn as ɵGetterFn, MethodFn as ɵMethodFn, SetterFn as ɵSetterFn} from './reflection/types';
 export {DirectRenderer as ɵDirectRenderer, RenderDebugInfo as ɵRenderDebugInfo} from './render/api';
+export {global as ɵglobal} from './util';
 export {makeDecorator as ɵmakeDecorator} from './util/decorators';
 export {isObservable as ɵisObservable, isPromise as ɵisPromise, merge as ɵmerge} from './util/lang';
