+++ modules/@angular/core/src/util/decorators.ts
@@ -6,11 +6,12 @@
  * found in the LICENSE file at https://angular.io/license
  */
 
-import {global, stringify} from '../facade/lang';
+import {stringify} from '../facade/lang';
 import {Type} from '../type';
+import {global} from '../util';
 
 let _nextClassId = 0;
-const Reflect = global.Reflect;
+const Reflect = global['Reflect'];
 
 /**
  * Declares the interface to be used with {@link Class}.
