+++ modules/@angular/core/test/testability/testability_spec.ts
@@ -12,7 +12,7 @@ import {Testability} from '@angular/core/src/testability/testability';
 import {NgZone} from '@angular/core/src/zone/ng_zone';
 import {AsyncTestCompleter, SpyObject, beforeEach, describe, expect, inject, it} from '@angular/core/testing/testing_internal';
 
-import {scheduleMicroTask} from '../../src/facade/lang';
+import {scheduleMicroTask} from '../../src/util';
