+++ modules/@angular/core/test/zone/ng_zone_spec.ts
@@ -10,7 +10,8 @@ import {NgZone} from '@angular/core/src/zone/ng_zone';
 import {async, fakeAsync, flushMicrotasks} from '@angular/core/testing';
 import {AsyncTestCompleter, Log, beforeEach, describe, expect, inject, it, xit} from '@angular/core/testing/testing_internal';
 import {browserDetection} from '@angular/platform-browser/testing/browser_util';
-import {isPresent, scheduleMicroTask} from '../../src/facade/lang';
+import {isPresent} from '../../src/facade/lang';
+import {scheduleMicroTask} from '../../src/util';
 
 const needsLongerTimers = browserDetection.isSlow || browserDetection.isEdge;
 const resultTimer = 1000;
